Add PUT requests

+++ flask_cities_api/api/routes.py
@@ -32,20 +32,34 @@ def cities_list():
         return jsonify({
             'count': pagination.total,
             'next': next,
+            'prev': prev,
             'posts': [city_serializer.dump(obj)
                       for obj in objs],
-            'prev': prev,
         })
 
     if request.method == 'POST':
-        name = request.json['name']
-        region_id = request.json['region_id']
 
-        new_object = City(name, region_id)
+        name = request.json.get('name', None)
+        region_id = request.json.get('region_id', None)
+
+        if not name:
+            return {"name": "Please provide a city name"}, 400
+
+        if not region_id:
+            return {"region_id": "Please provide a region_id"}, 400
+
+        try:
+            Region.query.get_or_404(region_id)
+        except:
+            return {"region_id": "Region with such id does not exist"}, 400
+
+        new_object = City(**request.json)
 
         db.session.add(new_object)
         db.session.commit()
-        return city_serializer.jsonify(new_object)
+
+        result = city_serializer.dump(new_object)
+        return jsonify(result)
 
 
 @api.route('/cities/<int:id>', methods=['GET', 'DELETE', 'PUT'])
@@ -59,10 +73,19 @@ def cities_detail(id):
         return jsonify(result)
 
     if request.method == 'PUT':
+        region_id = request.json['region_id']
+
+        try:
+            Region.query.get_or_404(region_id)
+        except:
+            return {"region_id": "Region with such id does not exist"}, 400
+
         obj.name = request.json['name']
-        obj.region_id = request.json['region_id']
+        obj.region_id = region_id
         obj.population = request.json['population']
         obj.year_founded = request.json['year_founded']
+
+        db.session.commit()
 
         result = city_serializer.dump(obj)
         return jsonify(result)
@@ -95,19 +118,24 @@ def regions_list():
         return jsonify({
             'count': pagination.total,
             'next': next,
+            'prev': prev,
             'posts': [region_serializer.dump(obj)
                       for obj in objs],
-            'prev': prev,
         })
 
     if request.method == 'POST':
-        name = request.json['name']
+        name = request.json.get('name', None)
+
+        if not name:
+            return {"name": "Please provide a region name"}, 400
 
         new_object = Region(name)
 
         db.session.add(new_object)
         db.session.commit()
-        return region_serializer.jsonify(new_object)
+
+        result = region_serializer.dump(new_object)
+        return jsonify(result)
 
 
 @api.route('/regions/<int:id>', methods=['GET', 'DELETE', 'PUT'])
@@ -120,10 +148,18 @@ def regions_detail(id):
         result = region_serializer.dump(obj)
         return jsonify(result)
 
-    # if request.method == 'PUT':
-    #     obj = City.query.get_or_404(id)
-    #     result = region_serializer.dump(obj)
-    #     return jsonify(result)
+    if request.method == 'PUT':
+
+        name = request.json.get('name', None)
+
+        if not name:
+            return {"name": "Please provide a region name"}, 400
+
+        obj.name = name
+        db.session.commit()
+
+        result = city_serializer.dump(obj)
+        return jsonify(result)
 
     if request.method == 'DELETE':
         db.session.delete(obj)
